+++ feature-libs/checkout/base/components/checkout-payment-method/checkout-payment-method.component.ts
@@ -27,6 +27,7 @@ import {
   PaymentDetails,
   TranslationService,
   UserPaymentService,
+  WindowRef,
 } from '@spartacus/core';
 import { Card, ICON_TYPE, SelectFocusUtility } from '@spartacus/storefront';
 import {
@@ -59,6 +60,7 @@ export class CheckoutPaymentMethodComponent implements OnInit, OnDestroy {
     optional: true,
   });
   @Optional() protected focusService = inject(SelectFocusUtility);
+  @Optional() protected windowRef = inject(WindowRef);
 
   cards$: Observable<{ content: Card; paymentMethod: PaymentDetails }[]>;
   iconTypes = ICON_TYPE;
@@ -220,7 +222,36 @@ export class CheckoutPaymentMethodComponent implements OnInit, OnDestroy {
 
     this.savePaymentMethod(paymentDetails);
     if (this.featureConfigService?.isEnabled('a11yFocusOnCardAfterSelecting')) {
-      this.focusService.focusCardAfterSelecting(this.isUpdating$);
+      this.focusCardAfterSelecting();
+    }
+  }
+
+  /**
+   * Restores the focus to the Card component after it has been selected and the checkout has finished updating.
+   * The focus is lost due to DOM changes making it otherwise impossible to target elements that have been removed.
+   */
+  focusCardAfterSelecting(): void {
+    const cardNodes = Array.from(
+      this.windowRef?.document.querySelectorAll('cx-card')
+    );
+    const triggeredCard =
+      this.windowRef?.document.activeElement?.closest('cx-card');
+
+    if (triggeredCard) {
+      const selectedCardIndex = cardNodes.indexOf(triggeredCard);
+      this.isUpdating$
+        .pipe(
+          filter((isUpdating) => !isUpdating),
+          take(1)
+        )
+        .subscribe(() => {
+          requestAnimationFrame(() => {
+            const selectedCard = this.windowRef?.document.querySelectorAll(
+              'cx-card'
+            )[selectedCardIndex] as HTMLElement;
+            this.focusService.findFirstFocusable(selectedCard)?.focus();
+          });
+        });
     }
   }
